Agregado de modelo Credenciales y modificación de Empleado en cada capa
Se modifica cada capa de empleados para reflejar la salida de las columnas de ruta de credencial y se agrega el modelo credenciales, modificación de métodos imprimir e imprimir frente de la clase EmpleadosController para mantener solo la ultima imagen de acuerdo a la vigencia

+++ controllers/EmpleadosController.php
@@ -7,6 +7,7 @@ use app\models\Empleados;
 use app\models\EmpleadosSearch;
 use app\models\Departamentos;
 use app\models\Encargados;        
+use app\models\Credenciales;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
@@ -78,6 +79,7 @@ class EmpleadosController extends Controller
                     'title'=> "Empleados #".$id,
                     'content'=>$this->renderAjax('view', [
                         'model' => $this->findModel($id),
+                        'ruta_credencial' => $this->findModel($id)->ruta_Credencial,
                     ]),
                     'footer'=> Html::button('Cerrar',['class'=>'btn btn-default pull-left','data-dismiss'=>"modal"]).
                             Html::a('Editar',['update','id'=>$id],['class'=>'btn btn-primary','role'=>'modal-remote'])
@@ -85,6 +87,7 @@ class EmpleadosController extends Controller
         }else{
             return $this->render('view', [
                 'model' => $this->findModel($id),
+                'ruta_credencial' => $this->findModel($id)->ruta_Credencial,
             ]);
         }
     }
@@ -119,7 +122,7 @@ class EmpleadosController extends Controller
                     'footer'=> Html::button('Cerrar',['class'=>'btn btn-default pull-left','data-dismiss'=>"modal"]).
                                 Html::button('Guardar',['class'=>'btn btn-primary','type'=>"submit"])
         
-                ];         
+                ];
             }else if($model->load($request->post())){
 
                 return $this->subirArchivos($model);
@@ -131,7 +134,7 @@ class EmpleadosController extends Controller
                 //             Html::a('Create More',['create'],['class'=>'btn btn-primary','role'=>'modal-remote'])
         
                 // ];  
-            }else{           
+            }else{
                 return [
                     'title'=> "Crear Nuevo Empleado",
                     'content'=>$this->renderAjax('create', [
@@ -142,7 +145,7 @@ class EmpleadosController extends Controller
                     'footer'=> Html::button('Cerrar',['class'=>'btn btn-default pull-left','data-dismiss'=>"modal"]).
                                 Html::button('Guardar',['class'=>'btn btn-primary','type'=>"submit"])
         
-                ];         
+                ];
             }
         }else{
             /*
@@ -279,7 +282,16 @@ class EmpleadosController extends Controller
         if ($dir) {
             $tmp = $model->ruta_firma;
             if (file_exists($tmp)) {
-                unlink($tmp);                
+                unlink($tmp);
+            }
+        }
+
+        $dir = Yii::$app->params['FotosEmpleados'];
+        
+        if ($dir) {
+            $tmp = $model->ruta_foto;
+            if (file_exists($tmp)) {
+                unlink($tmp);
             }
         }
 
@@ -299,21 +311,24 @@ class EmpleadosController extends Controller
         }
     }
 
-
     public function actionImprimir($id)
     {
-        $this->imprimirFrente($id);
+        // $ruta_credencial
+        // $model->ruta_Credencial;
+        $model = $this->findModel($id);
+        
+        if(isset($model)){
+            $this->imprimirFrente($model, $id, !empty($model->ruta_Credencial));
+        }
     }
 
-
-    public function imprimirFrente($id)
-    {   
-        $model = $this->findModel($id);
+    public function imprimirFrente($model, $id, $existeCredencial)
+    {
         $nombre = $model->nombre;
         $apellidos = $model->ap_paterno . ' ' . $model->ap_materno;
         $departamento = $model->departamento_nombre;
         $categoria = $model->categoria;
-        $tipo_contrato = "CONFIANZA";
+        $tipo_contrato = Yii::$app->params['TipoContrato'];
         $ruta_foto = $model->ruta_foto;
         $ruta_firma = $model->ruta_firma;
         $tam_extra_chico = Yii::$app->params['TamExtraChico'];
@@ -321,13 +336,17 @@ class EmpleadosController extends Controller
         $tam_mediano = Yii::$app->params['TamMediano'];
         $tam_grande = Yii::$app->params['TamGrande'];
         $tam_extra_grande = Yii::$app->params['TamExtraGrande'];
+        $inicio_vigencia = $model->fecha_inicio_vigencia;
+        $termino_vigencia = $model->fecha_termino_vigencia;
+        $ruta_credencial = $model->ruta_Credencial;
+        // $ruta_cred_v = $model->ruta_credencial_v;
 
-        $homedir = dirname(__DIR__, 1);                
+        $homedir = dirname(__DIR__, 1);
         $fuente1 = $homedir . Yii::$app->params['RutaFuentes'] . Yii::$app->params['Fuente1'];
         $fuente2 = $homedir . Yii::$app->params['RutaFuentes'] . Yii::$app->params['Fuente2'];
-        // $fuente1 = __DIR__ . "/" . "/web/fonts/"" . "sansita.ttf";        
+        // $fuente1 = __DIR__ . "/" . "/web/fonts/"" . "sansita.ttf";
         $imgPlantilla = $homedir . Yii::$app->params['RutaPlantillas'] . 'CRED GUBERNATURA F.png';
-        // $imgPlantilla = __DIR__ . "/" . "imagen.png";        
+        // $imgPlantilla = __DIR__ . "/" . "imagen.png";
         // $rutaCredenciales = Yii::$app->params['Credenciales'];
 
         $imgFoto = $homedir . '/web/' . $ruta_foto;
@@ -337,9 +356,9 @@ class EmpleadosController extends Controller
         // imagen => C:\repositorio\credencial\controllers/imagen.png
         // fuentes => C:\repositorio\credencial\controllers/sansita.ttf
         // $fuente1 = Yii::$app->params['Fuentes'] . "sansita.ttf";
-        // $imgPlantilla = $rutaPlantillas . 'CRED GUBERNATURA F.jpg';        
+        // $imgPlantilla = $rutaPlantillas . 'CRED GUBERNATURA F.jpg';
         // $imagen = imagecreatefrompng($imgPlantilla);
-        // $imagen = imagecreatefromjpeg($imgPlantilla);        
+        // $imagen = imagecreatefromjpeg($imgPlantilla);
 
         // Crear instancias de imágenes
         $plantilla = imagecreatefrompng($imgPlantilla);
@@ -387,12 +406,12 @@ class EmpleadosController extends Controller
         
         // $x4 = 90;
         $bbox = imageftbbox($tam_extra_chico, 0, $fuente2, $categoria);
-        $x4 = $bbox[0] + (imagesx($contenedor) / 2) - ($bbox[4] / 2) - 10;
+        $x4 = $bbox[0] + (imagesx($contenedor) / 2) - ($bbox[4] / 2) - 5;
         $y4 = $y3 + $espacio + $tam_extra_chico;
         
         // $x5 = 90;
         $bbox = imageftbbox($tam_mediano, 0, $fuente2, $tipo_contrato);
-        $x5 = $bbox[0] + (imagesx($contenedor) / 2) - ($bbox[4] / 2) - 15;
+        $x5 = $bbox[0] + (imagesx($contenedor) / 2) - ($bbox[4] / 2) - 10;
         $y5 = $y4 + $espacio + $tam_extra_chico;
 
         imagettftext($contenedor, $tam_grande, $angulo, $x, $y, $color1, $fuente1, $nombre);
@@ -400,6 +419,9 @@ class EmpleadosController extends Controller
         imagettftext($contenedor, $tam_extra_chico, $angulo, $x3, $y3, $color1, $fuente2, $departamento);
         imagettftext($contenedor, $tam_extra_chico, $angulo, $x4, $y4, $color1, $fuente2, $categoria);
         imagettftext($contenedor, $tam_mediano, $angulo, $x5, $y5, $color2, $fuente1, $tipo_contrato);
+
+        // obtener carpeta en la que se guardará
+        $carpeta = Utilidades::crearCarpetaCredencial(Yii::$app->params['Credenciales'], $inicio_vigencia, $termino_vigencia);
 
         // Imprimir y liberar memoria
         /* Version para visualizar en pantalla
@@ -419,29 +441,78 @@ class EmpleadosController extends Controller
         imagedestroy($contenedor);*/
 
         // Versión para guardar el archivo en servidor
+        $salida = $carpeta . 'cred_'.Utilidades::nombreArchivo().'.png';
+
+        $inicio_seg = strtotime($inicio_vigencia);
+        $termino_seg = strtotime($termino_vigencia);
+
+        $credenciales = '';
+
         header("Content-Type: image/png");
-        $salida = "archivos/credenciales/procesada_1.png";
         imagepng($contenedor, $salida);
         imagedestroy($contenedor);
 
-        /*$dir = Yii::$app->params['FirmasEmpleados'];
-        
-                if ($dir) {
-                    $tmp = $model->ruta_firma;
-                    if (file_exists($tmp)) {
-                        unlink($tmp);                
+        if(!$existeCredencial) {
+            $credenciales = new Credenciales();
+            $credenciales->id_empleado = $model->id;
+            $credenciales->anio_inicio = date("Y", $inicio_seg);
+            $credenciales->anio_termino = date("Y", $termino_seg);
+            $credenciales->ruta_credencial_f = $salida;
+            $credenciales->save();
+        } else {
+                // echo '<pre>carpeta:'; print_r($carpeta); echo '</pre>';
+                // echo '<pre>ruta_cred:'; print_r($model->ruta_Credencial); echo '</pre>';
+            
+            // echo '<pre>id_emp:'; print_r($model->id); echo '</pre>';
+            
+            if($model->id != null)
+            {
+                    // echo '<pre>inicio:'; print_r(date("Y", $inicio_seg)); echo '</pre>';
+                    // echo '<pre>termino:'; print_r(date("Y", $termino_seg)); echo '</pre>';
+                    
+                $credenciales = Credenciales::findOne([
+                    'id_empleado' => $model->id,
+                    'anio_inicio' => date("Y", $inicio_seg),
+                    'anio_termino' => date("Y", $termino_seg),
+                ]);
+
+                // echo '<pre>'; print_r($credenciales); echo '</pre>';
+                // echo '<pre>salida:'; print_r($salida); echo '</pre>';
+                if(strpos($credenciales->ruta_credencial_f, $carpeta) !== false){
+                // echo '<pre>carpeta:'; print_r($carpeta); echo '</pre>';
+                // echo '<pre>ruta_cred:'; print_r($credenciales->ruta_credencial_f); echo '</pre>';
+                
+                // echo '<pre>'; var_dump( strpos($credenciales->ruta_credencial_f, $carpeta)); echo '</pre>';
+                // die();
+                    if (file_exists($credenciales->ruta_credencial_f)) {
+                        unlink($credenciales->ruta_credencial_f);
                     }
                 }
 
-                $rutaArchivo = Yii::$app->params['FirmasEmpleados'].Utilidades::nombreArchivo().'.'.$model->archivo_firma->extension;*/
+                $credenciales->ruta_credencial_f = $salida;
+                $credenciales->save();
+            }
+                // actualizar la credencial
+                //**** $credenciales = 
+            
 
+            /*$dir = Yii::$app->params['FirmasEmpleados'];
+            
+                    if ($dir) {
+                        $tmp = $model->ruta_firma;
+                        if (file_exists($tmp)) {
+                            unlink($tmp);
+                        }
+                    }
 
+                    $rutaArchivo = Yii::$app->params['FirmasEmpleados'].Utilidades::nombreArchivo().'.'.$model->archivo_firma->extension;*/
+        }
     }
 
 
     public function imprimirAdverso($id)
     {
-        $homedir = dirname(__DIR__, 1);                
+        $homedir = dirname(__DIR__, 1);
         $fuente1 = $homedir . Yii::$app->params['RutaFuentes'] . Yii::$app->params['Fuente1'];
         // echo '<pre>'; print_r($fuente1); echo '</pre>';
         // die();
@@ -453,9 +524,9 @@ class EmpleadosController extends Controller
         // imagen => C:\repositorio\credencial\controllers/imagen.png
         // fuentes => C:\repositorio\credencial\controllers/sansita.ttf
         // $fuente1 = Yii::$app->params['Fuentes'] . "sansita.ttf";
-        // $nombreImagen = $rutaPlantillas . 'CRED GUBERNATURA F.jpg';        
+        // $nombreImagen = $rutaPlantillas . 'CRED GUBERNATURA F.jpg';
         // $imagen = imagecreatefrompng($nombreImagen);
-        $imagen = imagecreatefromjpeg($nombreImagen);        
+        $imagen = imagecreatefromjpeg($nombreImagen);
         $color = imagecolorallocate($imagen, 0, 0, 0);
         $texto1 = "texto1";
         $texto2 = "texto2";
@@ -505,7 +576,7 @@ class EmpleadosController extends Controller
                 if ($dir) {
                     $tmp = $model->ruta_firma;
                     if (file_exists($tmp)) {
-                        unlink($tmp);                
+                        unlink($tmp);
                     }
                 }
 
@@ -526,7 +597,7 @@ class EmpleadosController extends Controller
                 if ($dir) {
                     $tmp = $model->ruta_foto;
                     if (file_exists($tmp)) {
-                        unlink($tmp);                
+                        unlink($tmp);
                     }
                 }
